Slowly learning MVVM, moved controller code to `MapleSeedU.ViewModels.MainViewWindowModel` and improved thread safety
