+++ MapleSeedU/Views/MainWindow.xaml
@@ -5,7 +5,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:Controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
     xmlns:ViewModels="clr-namespace:MapleSeedU.ViewModels" x:Class="MapleSeedU.Views.MainWindow"
-    mc:Ignorable="d" Width="823" Height="552" ResizeMode="NoResize" WindowStartupLocation="CenterScreen" Icon="/MapleSeedU;component/Seed.ico" Title="Maple Seed Lite" IsMaxRestoreButtonEnabled="False" SaveWindowPosition="True" ShowMaxRestoreButton="False" BorderBrush="Black" Loaded="Window_Loaded" Closing="Window_Closing">
+    mc:Ignorable="d" Width="823" Height="552" ResizeMode="NoResize" WindowStartupLocation="CenterScreen" Icon="/MapleSeedU;component/Seed.ico" Title="Maple Seed Lite" IsMaxRestoreButtonEnabled="False" SaveWindowPosition="True" ShowMaxRestoreButton="False" BorderBrush="Black">
 
     <Controls:MetroWindow.DataContext>
         <ViewModels:MainWindowViewModel/>
